Processing split away from UI functions and threaded to make UI non-blocking. Prevent loaded datasets from overflowing off of page. Layout tweaks.

+++ corvolauncher/gui/job_runners/cellxgene_worker.py
@@ -6,26 +6,31 @@ from functools import partial
 from subprocess import Popen, DEVNULL, STDOUT
 
 import requests
-from PyQt5.QtCore import QRunnable, pyqtSlot, pyqtSignal, QObject, Qt
+from PyQt5.QtCore import QRunnable, pyqtSlot, pyqtSignal, QObject, Qt, QRect, QThread
+from PyQt5.QtGui import QImage
 from PyQt5.QtWidgets import QVBoxLayout, QFrame, QHBoxLayout, QPushButton, QLabel
 
 from corvolauncher.gui.job_runners.worker_signals import WorkerSignals
-from corvolauncher.utilities.cellxgene_JSON_requests import CellxGeneJSONRequests
+from corvolauncher.utilities.cellxgene_json_requests import CellxGeneJSONRequests
 
 
-class CellXGeneDownloadWorker(QRunnable):
+class DatasetDownloadWorker(QRunnable):
+    """
+    Attempts to download a dataset with a provided name and dictionary of corresponding IDs expected by the CellXGene
+    AWS hash generator. Prints server error message if encountered (not yet linked to error slot).
+        """
 
-    def __init__(self, parent, ids, name, unique_name):
-        super(CellXGeneDownloadWorker, self).__init__()
+    def __init__(self, parent, name, ids):
+        super(DatasetDownloadWorker, self).__init__()
 
         self.parent = parent
         self.ids = ids
         self.name = name
-        self.unique_name = unique_name
+
+        self.unique_name = self.parent.parent.parent.moniker_recursively(self.name + "_corvo_RAW.h5ad", "datasets")
 
         self.shutdown = False
         self.signals = WorkerSignals()
-
 
         # only takes 0.3s to init and makes worker much simpler to instantiate each download
 
@@ -44,11 +49,10 @@ class CellXGeneDownloadWorker(QRunnable):
                 print(aws_hash["detail"])
                 # can return json: {'detail': 'An internal server error has occurred. Please try again later.',
                 # 'status': 500, 'title': 'Internal Server Error', 'type': 'about:blank'}
+                # first checking that server error has not occurred
             except KeyError:
-                h5ad_dataset = requests.get(aws_hash["presigned_url"], stream=True)
                 # note some unusual characters cannot be parsed by the hdf5 reader
-
-                # h5ad_dataset = requests.get(aws_hash["presigned_url"], stream=True)
+                h5ad_dataset = requests.get(aws_hash["presigned_url"], stream=True)
                 total_length = h5ad_dataset.headers.get('content-length')
                 self.signals.file_size.emit(int(total_length))
                 if total_length is None:  # no content length header
@@ -66,9 +70,9 @@ class CellXGeneDownloadWorker(QRunnable):
                                 done = int(50 * dl / total_length)
                                 self.signals.progress.emit(done)
                             else:
-                                os.remove("../resources/datasets/" + self.unique_name)
                                 break
-
+                    if self.shutdown:  # place here as file is used by open process in loop
+                        os.remove("../resources/datasets/" + self.unique_name)
         except Exception:
             traceback.print_exc()
             exctype, value = sys.exc_info()[:2]
@@ -76,58 +80,40 @@ class CellXGeneDownloadWorker(QRunnable):
         else:
             self.signals.result.emit()
         finally:
+            print("emitting finished")
             self.signals.finished.emit()  # Done
 
     def stop(self):
+        print("stopping")
         self.shutdown = True
 
 
-class CellXGeneJSONWorker(QRunnable):
-    def __init__(self, parent, index, *args, **kwargs):
-        super(CellXGeneJSONWorker, self).__init__()
+class DatasetInfoWorker(QRunnable):
+    """
+    Fetch the dataset and filetype IDs for a given dataset located at a QComboBox index. Assumes the index and
+    dataset names exist.
+    """
 
+    def __init__(self, parent, index, *args, **kwargs):
+        super(DatasetInfoWorker, self).__init__()
+
+        self.parent = parent
+        self.index = index
         self.args = args
         self.kwargs = kwargs
         self.signals = WorkerSignals()
-        self.parent = parent
-        self.index = index
 
     @pyqtSlot()
     def run(self):
-        if self.index != 0:
+        try:
+            self.signals.running.emit()
             dataset_map = self.parent.data_fetcher.get_collection_datasets(
                 self.parent.collection_map[self.parent.collection_container.itemText(self.index)])
-
-            dset_button_layout = QVBoxLayout()
-            dset_button_layout.setSpacing(10)
-            dset_button_frame = QFrame()
-            dset_button_layout.setAlignment(Qt.AlignLeft)
-
-            for dset in dataset_map.keys():
-                dset_layout = QHBoxLayout()
-                dset_download_button = QPushButton("Download")
-                dset_download_button.setObjectName(dset)
-                dset_download_button.setFixedWidth(65)
-                dset_download_button.setFixedHeight(20)
-                dset_download_button.clicked.connect(partial(self.parent.launch_download, dset, dataset_map))
-                dset_layout.addWidget(dset_download_button)
-
-                # dset_cancel_button = QPushButton("Cancel")
-                # dset_cancel_button.setObjectName(dset + "cancel")
-                # dset_download_button.setFixedWidth(65)
-                # dset_cancel_button.clicked.connect(self.cancel)
-                # dset_cancel_button.hide()
-                # dset_layout.addWidget(dset_cancel_button)
-
-                dset_label = QLabel(dset)
-                dset_label.setMinimumHeight(30)
-                dset_label.setFixedWidth(self.parent.collection_container.width() - dset_download_button.width())
-                # dset_label.setFixedWidth(self.collection_container.width() - 125)
-                # dset_label.setMaximumWidth(self.collection_container.width() - dset_download_button.width())
-                dset_label.setWordWrap(True)
-                dset_layout.addWidget(dset_label)
-                dset_button_layout.addLayout(dset_layout)
-            dset_button_frame.setLayout(dset_button_layout)
-            self.parent.fetcher_layout.addWidget(dset_button_frame)
-        print("done")
-
+        except Exception:
+            traceback.print_exc()
+            exctype, value = sys.exc_info()[:2]
+            self.signals.error.emit((exctype, value, traceback.format_exc()))
+        else:
+            self.signals.dict_result.emit(dataset_map)
+        finally:
+            self.signals.finished.emit()  # Done
